% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\docType{data}
\name{data_options}
\alias{data_options}
\title{Available datasets}
\format{
An object of class \code{list} of length 3.
}
\usage{
data_options
}
\description{
Named list, with names being concisely named human-readable descriptions,
and values being short programming-friendly labels.
}
\keyword{datasets}
