% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_parse_string.R
\name{group_same_words}
\alias{group_same_words}
\title{returns a data frame of word: keyword pairs}
\usage{
group_same_words(list_of_words)
}
\arguments{
\item{list_of_words}{list, with each element being a vector of words.
see example output of \code{parse_string}}
}
\description{
returns a data frame of word: keyword pairs
}
