% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_quotes.R
\name{html_link}
\alias{html_link}
\title{Create html formatted url}
\usage{
html_link(url, display_text = "link")
}
\arguments{
\item{url}{url}

\item{display_text}{text to display, optionally HTML formatted}
}
\description{
Create html formatted url
}
