% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{load_data}
\alias{load_data}
\title{Load preprocessed data for word counting}
\usage{
load_data(datasource = c("marx_cap", "lenin_en", "lenin_ru"))
}
\arguments{
\item{datasource}{string indicating which data source to use}
}
\description{
Some data sources have metadata etc about the text (table of contents, etc).
This function removes unnecessary rows.
}
