% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_parse_string.R
\name{parse_string}
\alias{parse_string}
\title{Parse comma-separated string input into keyword vector}
\usage{
parse_string(word_input)
}
\arguments{
\item{word_input}{string of comma separated key words}
}
\description{
Parse comma-separated string input into keyword vector
}
