% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_quotes.R
\name{post_process_quotes}
\alias{post_process_quotes}
\title{Data source-specific parsing for display}
\usage{
post_process_quotes(counts, datasource = c("marx_cap", "lenin_en", "lenin_ru"))
}
\arguments{
\item{counts}{data frame after counting word mentions}

\item{datasource}{shortname of data source}
}
\description{
Data source-specific parsing for display
}
