% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_quotes.R
\name{url_leninature}
\alias{url_leninature}
\title{Full url to Marxists.org given leninature url}
\usage{
url_leninature(url)
}
\arguments{
\item{url}{abbreviated url requiring pre-fix}
}
\description{
Full url to Marxists.org given leninature url
}
