% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_quotes.R
\name{url_marx_cap}
\alias{url_marx_cap}
\title{url to Marxists.org given chapter of Capital}
\usage{
url_marx_cap(chapter)
}
\arguments{
\item{chapter}{chapter number (integer)}
}
\description{
url to Marxists.org given chapter of Capital
}
