% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlds-package.R
\docType{package}
\name{mlds-package}
\alias{mlds}
\alias{mlds-package}
\title{mlds: Explore the works of Marxist-Leninists via a Shiny app}
\description{
This Shiny app allows users to query Marxist-Leninist texts for keywords of interest, and provides relevant passages and word frequencies in the form of human-readable tables and graphs.
}
\author{
\strong{Maintainer}: Alice Circe Malone \email{circemalone@gmail.com}

}
\keyword{internal}
